07.27.2017
Additions:
I don't care where you go, but you can't stay here - Added the ability to remove badges from graduated and inactive users, but made it so that badges cannot be added to those users.
Check marks the spot - When users are editing the calendar, the semester they are viewing is now highlighted in the list of avaliable semester. Also changed the active semester to have a check by it.
Shiny! - User profiles in the student roster now have a glow around them when they are being hovered over

Changes:
Fake chieves - Fixed an issue where some users that should not have access to badges, had the ability to earn badges.
Dates are hard - Changed the date display of the standard user calendar from numbers to text (i.e. 07/27/2017 to July 27, 2017